+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Add admin GET survey responses API [#66](https://github.com/rokwire/surveys-building-block/issues/66)
 ### Added
 - Sort Order when showing all (public) surveys [#64](https://github.com/rokwire/surveys-building-block/issues/64)